[SYCL][NFC] Outline `vec::convert` helpers
Current `vec::convert` implementation is not efficient, because it does
per-element conversion even though there are SPIR-V instructions to do a
whole vector convert operation through a single instruction.

There is an upcoming improvement of `vec::convert` implementation to
benefit from native SPIR-V vector conversion instructions, but it will
require siginifcant revamp of the implementation, potentially making it
take more lines of code.

Therefore, to simplify further development, review and code maintanence,
all helpers related to `vec::convert` implementation were outlined into a
separate file by this commit.
